New data structure for better overview of modules
