+++ Stopwatch/Main.py
@@ -1,10 +1,6 @@
 import sys
 import time
 import threading
-import wave
-import serial
-import os
-import os.path
 from PyQt5 import uic
 from PyQt5.QtWidgets import *
 from PyQt5.uic.properties import QtWidgets
@@ -12,16 +8,17 @@ from PyQt5 import QtWidgets
 from PyQt5 import QtCore, QtGui
 from PyQt5.QtGui import QColor
 import numpy as np
-from pydub import AudioSegment
-import pyqtgraph as pg
 from pygame import mixer
-import serial.tools.list_ports as port_list
 import cProfile
 import pstats
+from ArduinoInterface import ArduinoInterface
+from Marker import MarkerList
+from AudioConverter import AudioConverter
 
 ui_file = 'C:\\Users\\felix\\PycharmProjects\\Controller\\Stopwatch\\stopwatch3.ui'
-backup_file = 'C:\\Users\\felix\\PycharmProjects\\Controller\\Stopwatch\\markers_backup.txt'
 audio_file_path = 'C:\\Users\\felix\\PycharmProjects\\Controller\\Stopwatch\\song.mp3'
+
+REDUCTION_FACTOR = 6 # Audio Sample reduction factor - plot
 
 receiver_ids = [1, 2, 3, 4, 5, 6]
 mode = 2  # when this program is used automatically Picture Mode (2) is used
@@ -29,87 +26,7 @@ brightness = 2  # set fixed global brightness for testing
 saturation = 0  # unused
 velocity = 0  # unused
 
-REDUCTION_FACTOR = 6
 
-RECEIVER_COUNT = 6
-
-
-class MarkerList:
-    def __init__(self):
-        self.List = []
-        if os.path.isfile(backup_file):
-            print("backup file found")
-            self.f = open(backup_file, "r")
-        else:
-            self.f = open(backup_file, "w")
-            print("no backup file found - new backup file created")
-        self.reload_backup()
-
-    def reload_backup(self):
-        self.f = open(backup_file, "r")
-        str_input = self.f.read()
-        backup_details = []
-        [backup_details.append(b.split()) for b in str_input.splitlines()]
-        print(backup_details)
-        self.List = backup_details
-        self.f.close()
-
-    def add_marker(self, number, time_stamp, ID, picture, send_status):
-        self.List.append([number, time_stamp, ID, picture, send_status])
-        self.update_backup_file()
-
-    def update_backup_file(self):
-        self.f = open(backup_file, "w")
-        for x in self.List:
-            for i in x:
-                self.f.write(str(i) + " ")
-            self.f.write("\n")
-        self.f.close()
-
-    def output_list_as_string(self):
-        output = ""
-        if len(self.List) > 0:
-            for x in self.List:
-                output += f"\nMarker {x[0]} - Time: {x[1]} - ID {x[2]} - Picture: {x[3]} - Status Sent: {x[4]}"
-        return output
-
-    def delete_last(self):
-        if len(self.List) > 0:
-            self.List.pop()
-            self.update_backup_file()
-            print("Last Marker deleted")
-        else:
-            print("List empty")
-
-    def reset_send_status_all(self):
-        for x in self.List:
-            x[4] = "False"
-
-    def get_marker(self):
-        return self.List
-
-    def get_marker_time_ms(self, index):
-        ms = 0
-        if index < len(self.List):
-            m = self.List[index]
-            ms = int(m[1][10]) * 10 + int(m[1][9]) * 100 + int(m[1][7]) * 1000 + int(m[1][6]) * 10000 + int(
-                m[1][4]) * 60 * 1000
-            if ms < 0 or ms > 1200000:  # 20 min
-                ms = 0
-        return ms
-
-    def set_marker_sent_status(self, marker_number, status_new):
-        for lm in self.List:
-            if lm[0] == marker_number:
-                lm[-1] = status_new
-                print("Set sent status of marker " + str(marker_number) + " to " + str(status_new))
-
-    def get_highest_marker_number(self):
-        if len(self.List) > 0:
-            # print("ListMarkers len: " + str(self.List[-1][0]))
-            return int(self.List[-1][0])
-        else:
-            return 0
 
 
 class MyGUI(QMainWindow):
@@ -131,7 +48,7 @@ class MyGUI(QMainWindow):
         self.paused = False
         self.plotwidget = 0
         self.sound_file = audio_file_path
-        self.arduino = ArduinoInterface()
+        self.arduino = ArduinoInterface(receiver_ids)
         self.arduino.find_ports()
         self.marker_list = MarkerList()
         self.audio_converter = AudioConverter()
@@ -241,14 +158,14 @@ class MyGUI(QMainWindow):
         self.table1.setRowCount(len(self.marker_list.List) + 1)
 
     def create_infos_table(self):
-        self.tableWidget.setRowCount(RECEIVER_COUNT)
-        for r in range(RECEIVER_COUNT):
+        self.tableWidget.setRowCount(len(receiver_ids))
+        for r in range(len(receiver_ids)):
             self.tableWidget.setItem(r, 0, QtWidgets.QTableWidgetItem("LED Poi " + str(r + 1)))
             self.tableWidget.setItem(r, 1, QtWidgets.QTableWidgetItem(str(self.arduino.voltages[r]) + " V"))
             self.tableWidget.setItem(r, 2, QtWidgets.QTableWidgetItem(str(self.arduino.signal_strength[r]) + " %"))
 
     def update_info_table(self):
-        for r in range(RECEIVER_COUNT):
+        for r in range(len(receiver_ids)):
             item = self.tableWidget.item(r, 1)
             item.setText(str(self.arduino.voltages[r]) + " V")
             item = self.tableWidget.item(r, 2)
@@ -387,103 +304,7 @@ class MyGUI(QMainWindow):
         sys.exit()
 
 
-class ArduinoInterface:
-    def __init__(self):
-        self.voltages = [0] * RECEIVER_COUNT
-        self.signal_strength = [0] * RECEIVER_COUNT
-        self.ports_names = []
-        self.ports_COMs = []
-        self.serialPort = None
 
-    def find_ports(self):
-        # Find connected Ports for Arduino
-        ports = list(port_list.comports())
-
-        for p in ports:
-            print(p)
-            self.ports_COMs.append(p[0])
-            self.ports_names.append(p[1])
-
-        index = 0
-        for p in self.ports_names:
-            index = index + 1
-            if "arduino uno" in p.lower():
-                self.serialPort = serial.Serial(
-                    port=self.ports_COMs[index - 1], baudrate=115200, bytesize=8, write_timeout=1, timeout=2,
-                    stopbits=serial.STOPBITS_ONE
-                )
-
-        if not self.serialPort.is_open: # open just the first port
-            self.serialPort = serial.Serial(
-                port=self.ports_COMs[0], baudrate=115200, bytesize=8, write_timeout=1, timeout=2,
-                stopbits=serial.STOPBITS_ONE
-            )
-
-    def go_all_black(self):
-        # Make all LEDs go black
-        for ids in receiver_ids:
-            self.send(mode, ids, 0, saturation, 0, velocity)
-
-    def print_receiver_infos(self):
-        for x in range(RECEIVER_COUNT):
-            print(f"Receiver {x} voltage: {self.voltages[x]} signal strength: {self.signal_strength[x]}")
-
-    # def send(self, channel, picturenum, brightness):
-    def send(self, mode, receiver_id, picture_hue, saturation, brightness_value, velocity):
-        starttime_SP_write = time.time_ns()
-        byte1, byte2, byte3, byte4, byte5, byte6 = int(mode), int(receiver_id), int(picture_hue), int(saturation), \
-                                                   int(brightness_value), int(velocity)
-
-        self.serialPort.write(chr(byte1).encode('latin_1') + chr(byte2).encode('latin_1') + chr(byte3).encode('latin_1') +
-                         chr(byte4).encode('latin_1') + chr(byte5).encode('latin_1') + chr(byte6).encode('latin_1'))
-        # print("Duration for serial port write " + str((time.time_ns() - starttime_SP_write) / 1000000) + " ms")
-        # self.serialPort.reset_input_buffer()  # Fixed some problems earlier!
-        # serialPort.reset_output_buffer()
-
-        if self.serialPort.in_waiting > 0:
-            # Read data out of the buffer until a carriage return / new line is found
-            bytearray = self.serialPort.readline()
-            res = str(bytearray.decode("Ascii"))
-            lst = res.split()
-            self.voltages = lst[:6]
-            self.signal_strength = lst[-6:]
-
-        print("Duration for serial port write and receive: " + str((time.time_ns() - starttime_SP_write) / 1000000) + " ms")
-            # self.print_receiver_infos()
-
-        # print("after readLine")
-
-        # Print the contents of the serial data
-        # try:
-        #     for d in res:
-        #         print(d, end="")
-        #     print("")
-        # except:
-        #     print("Error")
-        #     pass
-
-
-class AudioConverter:
-    def __init__(self):
-        self.starttime = 0
-
-    def convert_mp3_to_array(self, input_file):
-        print("Started WAV Conversion")
-        self.starttime = time.time()
-        sound = AudioSegment.from_mp3(input_file)
-        sound.export('converted.wav', format="wav")
-        with wave.open('converted.wav', 'r') as wav_file:
-            signal = wav_file.readframes(-1)
-            # signal = np.fromstring(signal, int)
-            signal = np.frombuffer(signal, int)
-            # Get time from indices
-            fs = wav_file.getframerate()
-            time_x = np.linspace(0, len(signal) / fs, num=int(len(signal)))
-        print("Conversion of " + str(input_file) + " done!")
-        duration = (int((time.time() - self.starttime) * 100)) / 100  # truncate to 2 decimals
-        print("Conversion took: " + str(duration) + " s")
-        print("Audio duration: " + str(int(len(signal) / fs / 60)) + ' min ' + str(int(len(signal) / fs % 60)) + ' s')
-        return time_x, signal
 
 
 def main():
